Alternate rendering optimizer

+++ src/main/java/cc/hyperium/mixins/profiler/MixinProfiler.java
@@ -52,9 +52,9 @@ public class MixinProfiler {
     private void endSection(CallbackInfo ci) {
         if (this.profilingEnabled) {
             long i = System.nanoTime();
-            long j = (Long) this.timestampList.get(this.timestampList.size() - 1);
+            long j = this.timestampList.get(this.timestampList.size() - 1);
             long k = i - j;
-
+           
             if (k > TimeUnit.MILLISECONDS.toNanos(20)) {
 //                System.out.println(System.currentTimeMillis() + " Something\'s taking too long! \'" + this.profilingSection + "\' took aprox " + (double) k / 1000000.0D + " ms");
 //                if (Minecraft.getMinecraft().renderGlobal == null) {
